% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dirs.R
\name{create_irp_dirs}
\alias{create_irp_dirs}
\title{Function to populate a clean project directory with R, Data, Documentation, Submission
and Correspondence folders.}
\usage{
create_irp_dirs(root_dir = here::here())
}
\arguments{
\item{root_dir}{defaults to here::here() the project root dir}
}
\description{
Please Note: Unless you run this function in the root project directory, you'll have to supply the path.
}
\examples{
# In a project directory
create_irp_dirs()
}
