% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_id.R
\name{encrypt_id}
\alias{encrypt_id}
\title{Encrypt IDs for U15 Projects}
\usage{
encrypt_id(id)
}
\arguments{
\item{id}{An id or vector of ids to encrypt}
}
\value{
an encrypted id
}
\description{
Encrypt IDs for U15 Projects
}
