% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_irp.R
\name{irp_path}
\alias{irp_path}
\title{Cross platform file path constructor for IRP Shared Drive}
\usage{
irp_path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors of director or file paths.}

\item{fsep}{the path separator to use (assumed to be ASCII).}
}
\description{
Cross platform file path constructor for IRP Shared Drive
}
