% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{queens_secondary}
\alias{queens_secondary}
\title{Queen's Secondary Colour Palettes}
\format{
An object of class \code{character} of length 11.
}
\usage{
queens_secondary
}
\description{
Queen's Secondary Colour Palettes
}
\keyword{datasets}
