% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{queens_tricolour}
\alias{queens_tricolour}
\title{Queen's Tricolour Palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
queens_tricolour
}
\description{
Queen's Tricolour Palette
}
\keyword{datasets}
