% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_credentials.R
\name{set_db_credentials}
\alias{set_db_credentials}
\title{Set Database UID and PWD in .Renviron file}
\usage{
set_db_credentials(UID = NA, PWD = NA)
}
\arguments{
\item{UID}{User ID to Set}

\item{PWD}{Password to Set}
}
\description{
Set Database UID and PWD in .Renviron file
}
